Added training session entity

+++ neuralserv-api/src/main/resources/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE number_image_entity (
     image_weights double precision[][]
 );
 CREATE TABLE training_session (
-    id serial primary key,
+    id bigint primary key,
     internal_training_size integer not null,
     external_training_size integer not null,
     internal_number_correct integer not null,
